+++ kiota_abstractions/authentication/anonymous_authentication_provider.py
@@ -1,3 +1,11 @@
+# ------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation.  All Rights Reserved.
+# Licensed under the MIT License.
+# See License in the project root for license information.
+# ------------------------------------------------------------------------------
+
+from typing import Any, Dict
+
 from ..request_information import RequestInformation
 from .authentication_provider import AuthenticationProvider
 
@@ -9,10 +17,15 @@ class AnonymousAuthenticationProvider(AuthenticationProvider):
         AuthenticationProvider (ABC): The abstract base class that this class implements
     """
 
-    async def authenticate_request(self, request: RequestInformation) -> None:
+    async def authenticate_request(
+        self,
+        request: RequestInformation,
+        additional_authentication_context: Dict[str, Any] = {}
+    ) -> None:
         """Authenticates the provided request information
 
         Args:
             request (RequestInformation): Request information object
+            additional_authentication_context (dict):
         """
         return